+++ app/src/main/java/de/passwordvault/view/passwords/activity_analysis/PasswordAnalysisScreen.kt
@@ -210,8 +210,6 @@ private fun ContentSection(
             0 -> GeneralTab(
                 securityScore = securityScore,
                 maxSecurityScore = maxSecurityScore,
-                thresholdGood = thresholdGood,
-                thresholdNeutral = thresholdNeutral,
                 numberOfWeakPasswords = weakPasswords.size,
                 numberOfIdenticalPasswords = numberOfIdenticalPasswords,
                 onWeakPasswordsClicked = {