+++ app/src/main/java/de/passwordvault/view/settings/activity_webview/WebViewScreen.kt
@@ -1,0 +1,124 @@
+package de.passwordvault.view.settings.activity_webview
+
+import android.view.ViewGroup
+import android.webkit.WebChromeClient
+import android.webkit.WebView
+import android.webkit.WebViewClient
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.core.view.ViewCompat
+import de.passwordvault.R
+
+
+/**
+ * Displays the screen which shows web pages within the app.
+ *
+ * @param viewModel     View model for the screen.
+ * @param onNavigateUp  Callback invoked to navigate up on the navigation stack.
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun WebViewScreen(
+    viewModel: WebViewViewModel,
+    onNavigateUp: () -> Unit
+) {
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(
+                        text = if(viewModel.title == null) { stringResource(R.string.app_name) } else { viewModel.title!! },
+                        color = MaterialTheme.colorScheme.onSurface,
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis
+                    )
+                },
+                navigationIcon = {
+                    IconButton(
+                        onClick = {
+                            onNavigateUp()
+                        }
+                    ) {
+                        Icon(
+                            painter = painterResource(R.drawable.ic_back),
+                            tint = MaterialTheme.colorScheme.onSurface,
+                            contentDescription = ""
+                        )
+                    }
+                },
+                colors = TopAppBarDefaults.topAppBarColors().copy(containerColor = MaterialTheme.colorScheme.surfaceContainerHighest),
+            )
+        }
+    ) { innerPadding ->
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier
+                .padding(innerPadding)
+                .fillMaxSize()
+        ) {
+            WebView(
+                url = viewModel.url,
+                client = MyWebViewClient(
+                    onPageFinished = { title ->
+                        viewModel.onPageFinished(title)
+                    }
+                )
+            )
+            if (viewModel.title == null) {
+                CircularProgressIndicator()
+            }
+        }
+    }
+}
+
+
+/**
+ * Composable displays a web view.
+ *
+ * @param url       URL to load within the web view.
+ * @param modifier  Modifier.
+ * @param client    Web view client.
+ */
+@Composable
+private fun WebView(
+    url: String,
+    modifier: Modifier = Modifier,
+    client: WebViewClient = WebViewClient()
+) {
+    //Unfortunately, there is no web view composable, so we need to use Android Views.
+    AndroidView(
+        modifier = modifier,
+        factory = { context ->
+            WebView(context).apply {
+                layoutParams = ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT)
+                settings.apply {
+                    loadsImagesAutomatically = true
+                }
+                webViewClient = client
+                webChromeClient = WebChromeClient()
+                ViewCompat.setNestedScrollingEnabled(this, true)
+                loadUrl(url)
+            }
+        },
+        update = { webView ->
+            webView.loadUrl(url)
+        }
+    )
+}